[output-mapping] TableStrategy::createSource methods moved into WorkspaceItemSourceFactory
