+++ libs/output-mapping/src/Writer/Table/Source/WorkspaceItemSourceFactoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\OutputMapping\Writer\Table\Source;
+
+interface WorkspaceItemSourceFactoryInterface
+{
+    public function createSource(string $sourcePathPrefix, string $sourceName): WorkspaceItemSource;
+}
